xapi: allow git-describe call to work in the container
The default for the Git fetcher is to use share history with the repo in
DL_DIR, but that cannot work when only the source tree is bind-mounted in
the container.

BB_GIT_NOSHARED forces git-clone to make a copy of the git history, so
git-describe can work.  A bitbake fix is necessary to also copy the
LFS objects so git-checkout does not fail in do_unpack.