create-install-image
2 recipes to build the initrd and the iso respectedly.

Those could be 2 tasks of a single recipe, but then do_build and do_deploy
need to be duplicated with non-standard naming.

FIXME: expecting separate repos for alma components, and generating a
single all-in-one repo, is causing issues.  If we change the first one
we're breaking support for upstream Alma repos.

So for now we're allowing fetching upstream deps from upstream servers.